% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmr.R
\name{lmr}
\alias{lmr}
\title{Custom Linear Regression Function}
\usage{
lmr(Y, X, data)
}
\arguments{
\item{Y}{Response which variable you focus on}

\item{X}{is the covariates which will affect Y}

\item{data}{Data frame containing the variables.}
}
\value{
A list containing the regression table and residuals summary.
}
\description{
This function performs a linear regression similar to lm() but without using the built-in lm function.
}
